Modernize codebase: automated ruff format and lint fixes across 321 files
Apply ruff check --fix and ruff format across the entire Python codebase.
This includes pyupgrade (UP) rules for modern Python 3.13+ syntax,
isort import ordering, unnecessary comprehension simplification,
type annotation modernization, and consistent code formatting.

Fix broken syntax in scripts/license_manager.py, scripts/test_disaster_recovery.py,
tests/test_extends.py, tests/test_security.py, and pbx/utils/tts.py caused by
malformed f-string expressions.

+++ pbx/utils/dtmf.py
@@ -52,7 +52,7 @@ class DTMFDetector:
         for freq in DTMF_LOW_FREQS + DTMF_HIGH_FREQS:
             # Calculate the nearest DFT bin for this frequency
             # 0.5 added for proper rounding to nearest integer
-            k = int(round(samples_per_frame * freq / sample_rate))
+            k = round(samples_per_frame * freq / sample_rate)
             omega = (2.0 * math.pi * k) / samples_per_frame
             self.coefficients[freq] = 2.0 * math.cos(omega)
 
@@ -303,7 +303,5 @@ class DTMFGenerator:
             samples.extend(tone_samples)
             samples.extend(gap_samples)
 
-        self.logger.info(
-            f"Generated DTMF sequence: {digits} ({len(samples)} samples)"
-        )
+        self.logger.info(f"Generated DTMF sequence: {digits} ({len(samples)} samples)")
         return samples
